+++ src/global/clack/setup-wizard-constants.js
@@ -1,0 +1,6 @@
+export class SetupWizardCancelledError extends Error {
+  constructor() {
+    super("Setup cancelled.");
+    this.name = "SetupWizardCancelledError";
+  }
+}